# exp: chunk preimage loop