--- a/src/poly/dcrt/sampler/trapdoor.rs
+++ b/src/poly/dcrt/sampler/trapdoor.rs
@@ -209,14 +209,9 @@ impl PolyTrapdoorSampler for DCRTPolyTrapdoorSampler {
             "Target matrix should have the same number of rows as the public matrix"
         );
 
-        // todo: real param and dummy param should have diff value
-        let chunk_size = 100;
         let num_block = target_cols.div_ceil(size);
         let k = params.modulus_bits();
-        debug_mem(format!(
-            "preimage before loop processing with chunksize {}, out of {}",
-            chunk_size, num_block
-        ));
+        debug_mem(format!("preimage before loop processing out of {}", num_block));
 
         let public_matrix = DCRTMatrixPtr::new_public_matrix(
             public_matrix,
@@ -228,27 +223,15 @@ impl PolyTrapdoorSampler for DCRTPolyTrapdoorSampler {
         );
 
         debug_mem("SetMatrixElement public_matrix_ptr completed");
-        let indices: Vec<_> = (0..num_block).collect();
-        let preimages: Vec<_> = parallel_chunk_iter!(indices, chunk_size)
-            .map(|chunk| {
-                chunk
-                    .iter()
-                    .map(|&i| {
-                        let start_col = i * size;
-                        let end_col = (start_col + size).min(target_cols);
-                        let target_block = target.slice(0, size, start_col, end_col);
-                        debug_mem(format!("preimage iter: start_col = {}", start_col));
-                        self.process_preimage_block(
-                            params,
-                            trapdoor,
-                            &public_matrix,
-                            &target_block,
-                            size,
-                        )
-                    })
-                    .collect::<Vec<_>>()
+
+        let preimages: Vec<_> = (0..num_block)
+            .map(|i| {
+                let start_col = i * size;
+                let end_col = (start_col + size).min(target_cols);
+                let target_block = target.slice(0, size, start_col, end_col);
+                debug_mem(format!("preimage iter: start_col = {}", start_col));
+                self.process_preimage_block(params, trapdoor, &public_matrix, &target_block, size)
             })
-            .flatten()
             .collect();
 
         log_mem("Collected preimages");